+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ### Fixed
-- Fixed a bug where it was possible to set conditions on a child blocks UI element, which would cause the UI element not to display
+- Fixed a bug where setting conditions on a child blocks UI element would cause the UI element not to display
 
 ## 3.7.6 - 2023-04-06